PSP: move internal types from extensions to policy.
