+++ pkg/apis/extensions/fuzzer/fuzzer.go
@@ -54,34 +54,6 @@ var Funcs = func(codecs runtimeserializer.CodecFactory) []interface{} {
 				j.RollingUpdate = &rollingUpdate
 			}
 		},
-		func(psp *extensions.PodSecurityPolicySpec, c fuzz.Continue) {
-			c.FuzzNoCustom(psp) // fuzz self without calling this function again
-
-			runAsUserRules := []extensions.RunAsUserStrategy{
-				extensions.RunAsUserStrategyMustRunAsNonRoot,
-				extensions.RunAsUserStrategyMustRunAs,
-				extensions.RunAsUserStrategyRunAsAny,
-			}
-			psp.RunAsUser.Rule = runAsUserRules[c.Rand.Intn(len(runAsUserRules))]
-
-			seLinuxRules := []extensions.SELinuxStrategy{
-				extensions.SELinuxStrategyMustRunAs,
-				extensions.SELinuxStrategyRunAsAny,
-			}
-			psp.SELinux.Rule = seLinuxRules[c.Rand.Intn(len(seLinuxRules))]
-
-			supplementalGroupsRules := []extensions.SupplementalGroupsStrategyType{
-				extensions.SupplementalGroupsStrategyRunAsAny,
-				extensions.SupplementalGroupsStrategyMustRunAs,
-			}
-			psp.SupplementalGroups.Rule = supplementalGroupsRules[c.Rand.Intn(len(supplementalGroupsRules))]
-
-			fsGroupRules := []extensions.FSGroupStrategyType{
-				extensions.FSGroupStrategyMustRunAs,
-				extensions.FSGroupStrategyRunAsAny,
-			}
-			psp.FSGroup.Rule = fsGroupRules[c.Rand.Intn(len(fsGroupRules))]
-		},
 		func(j *extensions.DaemonSetSpec, c fuzz.Continue) {
 			c.FuzzNoCustom(j) // fuzz self without calling this function again
 			rhl := int32(c.Rand.Int31())
