+++ pkg/apis/policy/fuzzer/fuzzer.go
@@ -30,5 +30,33 @@ var Funcs = func(codecs runtimeserializer.CodecFactory) []interface{} {
 			c.FuzzNoCustom(s) // fuzz self without calling this function again
 			s.PodDisruptionsAllowed = int32(c.Rand.Intn(2))
 		},
+		func(psp *policy.PodSecurityPolicySpec, c fuzz.Continue) {
+			c.FuzzNoCustom(psp) // fuzz self without calling this function again
+
+			runAsUserRules := []policy.RunAsUserStrategy{
+				policy.RunAsUserStrategyMustRunAsNonRoot,
+				policy.RunAsUserStrategyMustRunAs,
+				policy.RunAsUserStrategyRunAsAny,
+			}
+			psp.RunAsUser.Rule = runAsUserRules[c.Rand.Intn(len(runAsUserRules))]
+
+			seLinuxRules := []policy.SELinuxStrategy{
+				policy.SELinuxStrategyMustRunAs,
+				policy.SELinuxStrategyRunAsAny,
+			}
+			psp.SELinux.Rule = seLinuxRules[c.Rand.Intn(len(seLinuxRules))]
+
+			supplementalGroupsRules := []policy.SupplementalGroupsStrategyType{
+				policy.SupplementalGroupsStrategyRunAsAny,
+				policy.SupplementalGroupsStrategyMustRunAs,
+			}
+			psp.SupplementalGroups.Rule = supplementalGroupsRules[c.Rand.Intn(len(supplementalGroupsRules))]
+
+			fsGroupRules := []policy.FSGroupStrategyType{
+				policy.FSGroupStrategyMustRunAs,
+				policy.FSGroupStrategyRunAsAny,
+			}
+			psp.FSGroup.Rule = fsGroupRules[c.Rand.Intn(len(fsGroupRules))]
+		},
 	}
 }
