+++ pkg/security/podsecuritypolicy/group/mustrunas_test.go
@@ -20,19 +20,19 @@ import (
 	"strings"
 	"testing"
 
-	"k8s.io/kubernetes/pkg/apis/extensions"
+	"k8s.io/kubernetes/pkg/apis/policy"
 )
 
 func TestMustRunAsOptions(t *testing.T) {
 	tests := map[string]struct {
-		ranges []extensions.GroupIDRange
+		ranges []policy.GroupIDRange
 		pass   bool
 	}{
 		"empty": {
-			ranges: []extensions.GroupIDRange{},
+			ranges: []policy.GroupIDRange{},
 		},
 		"ranges": {
-			ranges: []extensions.GroupIDRange{
+			ranges: []policy.GroupIDRange{
 				{Min: 1, Max: 1},
 			},
 			pass: true,
@@ -52,23 +52,23 @@ func TestMustRunAsOptions(t *testing.T) {
 
 func TestGenerate(t *testing.T) {
 	tests := map[string]struct {
-		ranges   []extensions.GroupIDRange
+		ranges   []policy.GroupIDRange
 		expected []int64
 	}{
 		"multi value": {
-			ranges: []extensions.GroupIDRange{
+			ranges: []policy.GroupIDRange{
 				{Min: 1, Max: 2},
 			},
 			expected: []int64{1},
 		},
 		"single value": {
-			ranges: []extensions.GroupIDRange{
+			ranges: []policy.GroupIDRange{
 				{Min: 1, Max: 1},
 			},
 			expected: []int64{1},
 		},
 		"multi range": {
-			ranges: []extensions.GroupIDRange{
+			ranges: []policy.GroupIDRange{
 				{Min: 1, Max: 1},
 				{Min: 2, Max: 500},
 			},
@@ -110,25 +110,25 @@ func TestGenerate(t *testing.T) {
 
 func TestValidate(t *testing.T) {
 	tests := map[string]struct {
-		ranges        []extensions.GroupIDRange
+		ranges        []policy.GroupIDRange
 		groups        []int64
 		expectedError string
 	}{
 		"nil security context": {
-			ranges: []extensions.GroupIDRange{
+			ranges: []policy.GroupIDRange{
 				{Min: 1, Max: 3},
 			},
 			expectedError: "unable to validate empty groups against required ranges",
 		},
 		"empty groups": {
-			ranges: []extensions.GroupIDRange{
+			ranges: []policy.GroupIDRange{
 				{Min: 1, Max: 3},
 			},
 			expectedError: "unable to validate empty groups against required ranges",
 		},
 		"not in range": {
 			groups: []int64{5},
-			ranges: []extensions.GroupIDRange{
+			ranges: []policy.GroupIDRange{
 				{Min: 1, Max: 3},
 				{Min: 4, Max: 4},
 			},
@@ -136,25 +136,25 @@ func TestValidate(t *testing.T) {
 		},
 		"in range 1": {
 			groups: []int64{2},
-			ranges: []extensions.GroupIDRange{
+			ranges: []policy.GroupIDRange{
 				{Min: 1, Max: 3},
 			},
 		},
 		"in range boundary min": {
 			groups: []int64{1},
-			ranges: []extensions.GroupIDRange{
+			ranges: []policy.GroupIDRange{
 				{Min: 1, Max: 3},
 			},
 		},
 		"in range boundary max": {
 			groups: []int64{3},
-			ranges: []extensions.GroupIDRange{
+			ranges: []policy.GroupIDRange{
 				{Min: 1, Max: 3},
 			},
 		},
 		"singular range": {
 			groups: []int64{4},
-			ranges: []extensions.GroupIDRange{
+			ranges: []policy.GroupIDRange{
 				{Min: 4, Max: 4},
 			},
 		},
