+++ pkg/security/podsecuritypolicy/user/mustrunas_test.go
@@ -18,14 +18,14 @@ package user
 
 import (
 	api "k8s.io/kubernetes/pkg/apis/core"
-	"k8s.io/kubernetes/pkg/apis/extensions"
+	"k8s.io/kubernetes/pkg/apis/policy"
 	"strings"
 	"testing"
 )
 
 func TestNewMustRunAs(t *testing.T) {
 	tests := map[string]struct {
-		opts *extensions.RunAsUserStrategyOptions
+		opts *policy.RunAsUserStrategyOptions
 		pass bool
 	}{
 		"nil opts": {
@@ -33,12 +33,12 @@ func TestNewMustRunAs(t *testing.T) {
 			pass: false,
 		},
 		"invalid opts": {
-			opts: &extensions.RunAsUserStrategyOptions{},
+			opts: &policy.RunAsUserStrategyOptions{},
 			pass: false,
 		},
 		"valid opts": {
-			opts: &extensions.RunAsUserStrategyOptions{
-				Ranges: []extensions.UserIDRange{
+			opts: &policy.RunAsUserStrategyOptions{
+				Ranges: []policy.UserIDRange{
 					{Min: 1, Max: 1},
 				},
 			},
@@ -57,8 +57,8 @@ func TestNewMustRunAs(t *testing.T) {
 }
 
 func TestGenerate(t *testing.T) {
-	opts := &extensions.RunAsUserStrategyOptions{
-		Ranges: []extensions.UserIDRange{
+	opts := &policy.RunAsUserStrategyOptions{
+		Ranges: []policy.UserIDRange{
 			{Min: 1, Max: 1},
 		},
 	}
@@ -76,8 +76,8 @@ func TestGenerate(t *testing.T) {
 }
 
 func TestValidate(t *testing.T) {
-	opts := &extensions.RunAsUserStrategyOptions{
-		Ranges: []extensions.UserIDRange{
+	opts := &policy.RunAsUserStrategyOptions{
+		Ranges: []policy.UserIDRange{
 			{Min: 1, Max: 1},
 			{Min: 10, Max: 20},
 		},
