+++ pkg/security/podsecuritypolicy/util/util.go
@@ -22,7 +22,7 @@ import (
 
 	"k8s.io/apimachinery/pkg/util/sets"
 	api "k8s.io/kubernetes/pkg/apis/core"
-	"k8s.io/kubernetes/pkg/apis/extensions"
+	"k8s.io/kubernetes/pkg/apis/policy"
 )
 
 const (
@@ -40,102 +40,102 @@ func GetAllFSTypesExcept(exceptions ...string) sets.String {
 func GetAllFSTypesAsSet() sets.String {
 	fstypes := sets.NewString()
 	fstypes.Insert(
-		string(extensions.HostPath),
-		string(extensions.AzureFile),
-		string(extensions.Flocker),
-		string(extensions.FlexVolume),
-		string(extensions.EmptyDir),
-		string(extensions.GCEPersistentDisk),
-		string(extensions.AWSElasticBlockStore),
-		string(extensions.GitRepo),
-		string(extensions.Secret),
-		string(extensions.NFS),
-		string(extensions.ISCSI),
-		string(extensions.Glusterfs),
-		string(extensions.PersistentVolumeClaim),
-		string(extensions.RBD),
-		string(extensions.Cinder),
-		string(extensions.CephFS),
-		string(extensions.DownwardAPI),
-		string(extensions.FC),
-		string(extensions.ConfigMap),
-		string(extensions.VsphereVolume),
-		string(extensions.Quobyte),
-		string(extensions.AzureDisk),
-		string(extensions.PhotonPersistentDisk),
-		string(extensions.StorageOS),
-		string(extensions.Projected),
-		string(extensions.PortworxVolume),
-		string(extensions.ScaleIO),
-		string(extensions.CSI),
+		string(policy.HostPath),
+		string(policy.AzureFile),
+		string(policy.Flocker),
+		string(policy.FlexVolume),
+		string(policy.EmptyDir),
+		string(policy.GCEPersistentDisk),
+		string(policy.AWSElasticBlockStore),
+		string(policy.GitRepo),
+		string(policy.Secret),
+		string(policy.NFS),
+		string(policy.ISCSI),
+		string(policy.Glusterfs),
+		string(policy.PersistentVolumeClaim),
+		string(policy.RBD),
+		string(policy.Cinder),
+		string(policy.CephFS),
+		string(policy.DownwardAPI),
+		string(policy.FC),
+		string(policy.ConfigMap),
+		string(policy.VsphereVolume),
+		string(policy.Quobyte),
+		string(policy.AzureDisk),
+		string(policy.PhotonPersistentDisk),
+		string(policy.StorageOS),
+		string(policy.Projected),
+		string(policy.PortworxVolume),
+		string(policy.ScaleIO),
+		string(policy.CSI),
 	)
 	return fstypes
 }
 
 // getVolumeFSType gets the FSType for a volume.
-func GetVolumeFSType(v api.Volume) (extensions.FSType, error) {
+func GetVolumeFSType(v api.Volume) (policy.FSType, error) {
 	switch {
 	case v.HostPath != nil:
-		return extensions.HostPath, nil
+		return policy.HostPath, nil
 	case v.EmptyDir != nil:
-		return extensions.EmptyDir, nil
+		return policy.EmptyDir, nil
 	case v.GCEPersistentDisk != nil:
-		return extensions.GCEPersistentDisk, nil
+		return policy.GCEPersistentDisk, nil
 	case v.AWSElasticBlockStore != nil:
-		return extensions.AWSElasticBlockStore, nil
+		return policy.AWSElasticBlockStore, nil
 	case v.GitRepo != nil:
-		return extensions.GitRepo, nil
+		return policy.GitRepo, nil
 	case v.Secret != nil:
-		return extensions.Secret, nil
+		return policy.Secret, nil
 	case v.NFS != nil:
-		return extensions.NFS, nil
+		return policy.NFS, nil
 	case v.ISCSI != nil:
-		return extensions.ISCSI, nil
+		return policy.ISCSI, nil
 	case v.Glusterfs != nil:
-		return extensions.Glusterfs, nil
+		return policy.Glusterfs, nil
 	case v.PersistentVolumeClaim != nil:
-		return extensions.PersistentVolumeClaim, nil
+		return policy.PersistentVolumeClaim, nil
 	case v.RBD != nil:
-		return extensions.RBD, nil
+		return policy.RBD, nil
 	case v.FlexVolume != nil:
-		return extensions.FlexVolume, nil
+		return policy.FlexVolume, nil
 	case v.Cinder != nil:
-		return extensions.Cinder, nil
+		return policy.Cinder, nil
 	case v.CephFS != nil:
-		return extensions.CephFS, nil
+		return policy.CephFS, nil
 	case v.Flocker != nil:
-		return extensions.Flocker, nil
+		return policy.Flocker, nil
 	case v.DownwardAPI != nil:
-		return extensions.DownwardAPI, nil
+		return policy.DownwardAPI, nil
 	case v.FC != nil:
-		return extensions.FC, nil
+		return policy.FC, nil
 	case v.AzureFile != nil:
-		return extensions.AzureFile, nil
+		return policy.AzureFile, nil
 	case v.ConfigMap != nil:
-		return extensions.ConfigMap, nil
+		return policy.ConfigMap, nil
 	case v.VsphereVolume != nil:
-		return extensions.VsphereVolume, nil
+		return policy.VsphereVolume, nil
 	case v.Quobyte != nil:
-		return extensions.Quobyte, nil
+		return policy.Quobyte, nil
 	case v.AzureDisk != nil:
-		return extensions.AzureDisk, nil
+		return policy.AzureDisk, nil
 	case v.PhotonPersistentDisk != nil:
-		return extensions.PhotonPersistentDisk, nil
+		return policy.PhotonPersistentDisk, nil
 	case v.StorageOS != nil:
-		return extensions.StorageOS, nil
+		return policy.StorageOS, nil
 	case v.Projected != nil:
-		return extensions.Projected, nil
+		return policy.Projected, nil
 	case v.PortworxVolume != nil:
-		return extensions.PortworxVolume, nil
+		return policy.PortworxVolume, nil
 	case v.ScaleIO != nil:
-		return extensions.ScaleIO, nil
+		return policy.ScaleIO, nil
 	}
 
 	return "", fmt.Errorf("unknown volume type for volume: %#v", v)
 }
 
 // FSTypeToStringSet converts an FSType slice to a string set.
-func FSTypeToStringSet(fsTypes []extensions.FSType) sets.String {
+func FSTypeToStringSet(fsTypes []policy.FSType) sets.String {
 	set := sets.NewString()
 	for _, v := range fsTypes {
 		set.Insert(string(v))
@@ -144,19 +144,19 @@ func FSTypeToStringSet(fsTypes []extensions.FSType) sets.String {
 }
 
 // PSPAllowsAllVolumes checks for FSTypeAll in the psp's allowed volumes.
-func PSPAllowsAllVolumes(psp *extensions.PodSecurityPolicy) bool {
-	return PSPAllowsFSType(psp, extensions.All)
+func PSPAllowsAllVolumes(psp *policy.PodSecurityPolicy) bool {
+	return PSPAllowsFSType(psp, policy.All)
 }
 
 // PSPAllowsFSType is a utility for checking if a PSP allows a particular FSType.
 // If all volumes are allowed then this will return true for any FSType passed.
-func PSPAllowsFSType(psp *extensions.PodSecurityPolicy, fsType extensions.FSType) bool {
+func PSPAllowsFSType(psp *policy.PodSecurityPolicy, fsType policy.FSType) bool {
 	if psp == nil {
 		return false
 	}
 
 	for _, v := range psp.Spec.Volumes {
-		if v == fsType || v == extensions.All {
+		if v == fsType || v == policy.All {
 			return true
 		}
 	}
@@ -164,18 +164,18 @@ func PSPAllowsFSType(psp *extensions.PodSecurityPolicy, fsType extensions.FSType
 }
 
 // UserFallsInRange is a utility to determine it the id falls in the valid range.
-func UserFallsInRange(id int64, rng extensions.UserIDRange) bool {
+func UserFallsInRange(id int64, rng policy.UserIDRange) bool {
 	return id >= rng.Min && id <= rng.Max
 }
 
 // GroupFallsInRange is a utility to determine it the id falls in the valid range.
-func GroupFallsInRange(id int64, rng extensions.GroupIDRange) bool {
+func GroupFallsInRange(id int64, rng policy.GroupIDRange) bool {
 	return id >= rng.Min && id <= rng.Max
 }
 
 // AllowsHostVolumePath is a utility for checking if a PSP allows the host volume path.
 // This only checks the path. You should still check to make sure the host volume fs type is allowed.
-func AllowsHostVolumePath(psp *extensions.PodSecurityPolicy, hostPath string) bool {
+func AllowsHostVolumePath(psp *policy.PodSecurityPolicy, hostPath string) bool {
 	if psp == nil {
 		return false
 	}
