+++ plugin/pkg/admission/security/podsecuritypolicy/admission_test.go
@@ -36,7 +36,6 @@ import (
 	"k8s.io/kubernetes/pkg/api/legacyscheme"
 	kapi "k8s.io/kubernetes/pkg/apis/core"
 	"k8s.io/kubernetes/pkg/apis/core/helper"
-	"k8s.io/kubernetes/pkg/apis/extensions"
 	"k8s.io/kubernetes/pkg/apis/policy"
 	informers "k8s.io/kubernetes/pkg/client/informers/informers_generated/internalversion"
 	"k8s.io/kubernetes/pkg/controller"
@@ -50,13 +49,13 @@ import (
 const defaultContainerName = "test-c"
 
 // NewTestAdmission provides an admission plugin with test implementations of internal structs.
-func NewTestAdmission(psps []*extensions.PodSecurityPolicy, authz authorizer.Authorizer) *PodSecurityPolicyPlugin {
+func NewTestAdmission(psps []*policy.PodSecurityPolicy, authz authorizer.Authorizer) *PodSecurityPolicyPlugin {
 	informerFactory := informers.NewSharedInformerFactory(nil, controller.NoResyncPeriodFunc())
-	store := informerFactory.Extensions().InternalVersion().PodSecurityPolicies().Informer().GetStore()
+	store := informerFactory.Policy().InternalVersion().PodSecurityPolicies().Informer().GetStore()
 	for _, psp := range psps {
 		store.Add(psp)
 	}
-	lister := informerFactory.Extensions().InternalVersion().PodSecurityPolicies().Lister()
+	lister := informerFactory.Policy().InternalVersion().PodSecurityPolicies().Lister()
 	if authz == nil {
 		authz = &TestAuthorizer{}
 	}
@@ -217,7 +216,7 @@ func TestAdmitSeccomp(t *testing.T) {
 				},
 			},
 		}
-		testPSPAdmit(k, []*extensions.PodSecurityPolicy{psp}, pod, v.shouldPassAdmit, v.shouldPassValidate, psp.Name, t)
+		testPSPAdmit(k, []*policy.PodSecurityPolicy{psp}, pod, v.shouldPassAdmit, v.shouldPassValidate, psp.Name, t)
 	}
 }
 
@@ -241,7 +240,7 @@ func TestAdmitPrivileged(t *testing.T) {
 
 	tests := map[string]struct {
 		pod                *kapi.Pod
-		psps               []*extensions.PodSecurityPolicy
+		psps               []*policy.PodSecurityPolicy
 		shouldPassAdmit    bool
 		shouldPassValidate bool
 		expectedPriv       *bool
@@ -249,7 +248,7 @@ func TestAdmitPrivileged(t *testing.T) {
 	}{
 		"pod with priv=nil allowed under non priv PSP": {
 			pod:                goodPod(),
-			psps:               []*extensions.PodSecurityPolicy{nonPrivilegedPSP},
+			psps:               []*policy.PodSecurityPolicy{nonPrivilegedPSP},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPriv:       nil,
@@ -257,7 +256,7 @@ func TestAdmitPrivileged(t *testing.T) {
 		},
 		"pod with priv=nil allowed under priv PSP": {
 			pod:                goodPod(),
-			psps:               []*extensions.PodSecurityPolicy{privilegedPSP},
+			psps:               []*policy.PodSecurityPolicy{privilegedPSP},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPriv:       nil,
@@ -265,7 +264,7 @@ func TestAdmitPrivileged(t *testing.T) {
 		},
 		"pod with priv=false allowed under non priv PSP": {
 			pod:                createPodWithPriv(false),
-			psps:               []*extensions.PodSecurityPolicy{nonPrivilegedPSP},
+			psps:               []*policy.PodSecurityPolicy{nonPrivilegedPSP},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPriv:       &falseValue,
@@ -273,7 +272,7 @@ func TestAdmitPrivileged(t *testing.T) {
 		},
 		"pod with priv=false allowed under priv PSP": {
 			pod:                createPodWithPriv(false),
-			psps:               []*extensions.PodSecurityPolicy{privilegedPSP},
+			psps:               []*policy.PodSecurityPolicy{privilegedPSP},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPriv:       &falseValue,
@@ -281,13 +280,13 @@ func TestAdmitPrivileged(t *testing.T) {
 		},
 		"pod with priv=true denied by non priv PSP": {
 			pod:                createPodWithPriv(true),
-			psps:               []*extensions.PodSecurityPolicy{nonPrivilegedPSP},
+			psps:               []*policy.PodSecurityPolicy{nonPrivilegedPSP},
 			shouldPassAdmit:    false,
 			shouldPassValidate: false,
 		},
 		"pod with priv=true allowed by priv PSP": {
 			pod:                createPodWithPriv(true),
-			psps:               []*extensions.PodSecurityPolicy{nonPrivilegedPSP, privilegedPSP},
+			psps:               []*policy.PodSecurityPolicy{nonPrivilegedPSP, privilegedPSP},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPriv:       &trueValue,
@@ -325,11 +324,11 @@ func defaultPod(t *testing.T, pod *kapi.Pod) *kapi.Pod {
 func TestAdmitPreferNonmutating(t *testing.T) {
 	mutating1 := restrictivePSP()
 	mutating1.Name = "mutating1"
-	mutating1.Spec.RunAsUser.Ranges = []extensions.UserIDRange{{Min: int64(1), Max: int64(1)}}
+	mutating1.Spec.RunAsUser.Ranges = []policy.UserIDRange{{Min: int64(1), Max: int64(1)}}
 
 	mutating2 := restrictivePSP()
 	mutating2.Name = "mutating2"
-	mutating2.Spec.RunAsUser.Ranges = []extensions.UserIDRange{{Min: int64(2), Max: int64(2)}}
+	mutating2.Spec.RunAsUser.Ranges = []policy.UserIDRange{{Min: int64(2), Max: int64(2)}}
 
 	privilegedPSP := permissivePSP()
 	privilegedPSP.Name = "privileged"
@@ -364,7 +363,7 @@ func TestAdmitPreferNonmutating(t *testing.T) {
 		operation             kadmission.Operation
 		pod                   *kapi.Pod
 		podBeforeUpdate       *kapi.Pod
-		psps                  []*extensions.PodSecurityPolicy
+		psps                  []*policy.PodSecurityPolicy
 		shouldPassValidate    bool
 		expectMutation        bool
 		expectedContainerUser *int64
@@ -373,7 +372,7 @@ func TestAdmitPreferNonmutating(t *testing.T) {
 		"pod should not be mutated by allow-all strategies": {
 			operation:             kadmission.Create,
 			pod:                   unprivilegedRunAsAnyPod.DeepCopy(),
-			psps:                  []*extensions.PodSecurityPolicy{privilegedPSP},
+			psps:                  []*policy.PodSecurityPolicy{privilegedPSP},
 			shouldPassValidate:    true,
 			expectMutation:        false,
 			expectedContainerUser: nil,
@@ -382,7 +381,7 @@ func TestAdmitPreferNonmutating(t *testing.T) {
 		"pod should prefer non-mutating PSP on create": {
 			operation:             kadmission.Create,
 			pod:                   unprivilegedRunAsAnyPod.DeepCopy(),
-			psps:                  []*extensions.PodSecurityPolicy{mutating2, mutating1, privilegedPSP},
+			psps:                  []*policy.PodSecurityPolicy{mutating2, mutating1, privilegedPSP},
 			shouldPassValidate:    true,
 			expectMutation:        false,
 			expectedContainerUser: nil,
@@ -391,7 +390,7 @@ func TestAdmitPreferNonmutating(t *testing.T) {
 		"pod should use deterministic mutating PSP on create": {
 			operation:             kadmission.Create,
 			pod:                   unprivilegedRunAsAnyPod.DeepCopy(),
-			psps:                  []*extensions.PodSecurityPolicy{mutating2, mutating1},
+			psps:                  []*policy.PodSecurityPolicy{mutating2, mutating1},
 			shouldPassValidate:    true,
 			expectMutation:        true,
 			expectedContainerUser: &mutating1.Spec.RunAsUser.Ranges[0].Min,
@@ -400,7 +399,7 @@ func TestAdmitPreferNonmutating(t *testing.T) {
 		"pod should use deterministic mutating PSP on create even if ValidatedPSPAnnotation is set": {
 			operation:             kadmission.Create,
 			pod:                   podWithAnnotation,
-			psps:                  []*extensions.PodSecurityPolicy{mutating2, mutating1},
+			psps:                  []*policy.PodSecurityPolicy{mutating2, mutating1},
 			shouldPassValidate:    true,
 			expectMutation:        true,
 			expectedContainerUser: &mutating1.Spec.RunAsUser.Ranges[0].Min,
@@ -410,7 +409,7 @@ func TestAdmitPreferNonmutating(t *testing.T) {
 			operation:             kadmission.Update,
 			pod:                   changedPodWithSC.DeepCopy(),
 			podBeforeUpdate:       podWithSC.DeepCopy(),
-			psps:                  []*extensions.PodSecurityPolicy{mutating2, mutating1, privilegedPSP},
+			psps:                  []*policy.PodSecurityPolicy{mutating2, mutating1, privilegedPSP},
 			shouldPassValidate:    true,
 			expectMutation:        false,
 			expectedContainerUser: nil,
@@ -420,7 +419,7 @@ func TestAdmitPreferNonmutating(t *testing.T) {
 			operation:             kadmission.Update,
 			pod:                   changedPod.DeepCopy(),
 			podBeforeUpdate:       unprivilegedRunAsAnyPod.DeepCopy(),
-			psps:                  []*extensions.PodSecurityPolicy{mutating2, mutating1},
+			psps:                  []*policy.PodSecurityPolicy{mutating2, mutating1},
 			shouldPassValidate:    false,
 			expectMutation:        false,
 			expectedContainerUser: nil,
@@ -430,7 +429,7 @@ func TestAdmitPreferNonmutating(t *testing.T) {
 			operation:             kadmission.Update,
 			pod:                   unprivilegedRunAsAnyPod.DeepCopy(),
 			podBeforeUpdate:       unprivilegedRunAsAnyPod.DeepCopy(),
-			psps:                  []*extensions.PodSecurityPolicy{mutating2, mutating1},
+			psps:                  []*policy.PodSecurityPolicy{mutating2, mutating1},
 			shouldPassValidate:    true,
 			expectMutation:        false,
 			expectedContainerUser: nil,
@@ -440,7 +439,7 @@ func TestAdmitPreferNonmutating(t *testing.T) {
 			operation:             kadmission.Update,
 			pod:                   gcChangedPod.DeepCopy(),
 			podBeforeUpdate:       unprivilegedRunAsAnyPod.DeepCopy(),
-			psps:                  []*extensions.PodSecurityPolicy{mutating2, mutating1},
+			psps:                  []*policy.PodSecurityPolicy{mutating2, mutating1},
 			shouldPassValidate:    true,
 			expectMutation:        false,
 			expectedContainerUser: nil,
@@ -516,11 +515,11 @@ func TestAdmitCaps(t *testing.T) {
 
 	allowAllInAllowed := restrictivePSP()
 	allowAllInAllowed.Name = "allowAllCapsInAllowed"
-	allowAllInAllowed.Spec.AllowedCapabilities = []kapi.Capability{extensions.AllowAllCapabilities}
+	allowAllInAllowed.Spec.AllowedCapabilities = []kapi.Capability{policy.AllowAllCapabilities}
 
 	tc := map[string]struct {
 		pod                  *kapi.Pod
-		psps                 []*extensions.PodSecurityPolicy
+		psps                 []*policy.PodSecurityPolicy
 		shouldPassAdmit      bool
 		shouldPassValidate   bool
 		expectedCapabilities *kapi.Capabilities
@@ -530,7 +529,7 @@ func TestAdmitCaps(t *testing.T) {
 		// should be rejected.
 		"should reject cap add when not allowed or required": {
 			pod:                createPodWithCaps(&kapi.Capabilities{Add: []kapi.Capability{"foo"}}),
-			psps:               []*extensions.PodSecurityPolicy{restricted},
+			psps:               []*policy.PodSecurityPolicy{restricted},
 			shouldPassAdmit:    false,
 			shouldPassValidate: false,
 		},
@@ -538,7 +537,7 @@ func TestAdmitCaps(t *testing.T) {
 		// to add the cap.
 		"should accept cap add when in allowed": {
 			pod:                createPodWithCaps(&kapi.Capabilities{Add: []kapi.Capability{"foo"}}),
-			psps:               []*extensions.PodSecurityPolicy{restricted, allowsFooInAllowed},
+			psps:               []*policy.PodSecurityPolicy{restricted, allowsFooInAllowed},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPSP:        allowsFooInAllowed.Name,
@@ -547,7 +546,7 @@ func TestAdmitCaps(t *testing.T) {
 		// to add the cap.
 		"should accept cap add when in required": {
 			pod:                createPodWithCaps(&kapi.Capabilities{Add: []kapi.Capability{"foo"}}),
-			psps:               []*extensions.PodSecurityPolicy{restricted, allowsFooInRequired},
+			psps:               []*policy.PodSecurityPolicy{restricted, allowsFooInRequired},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPSP:        allowsFooInRequired.Name,
@@ -556,7 +555,7 @@ func TestAdmitCaps(t *testing.T) {
 		// in the verification of adds and verification of drops
 		"should reject cap add when requested cap is required to be dropped": {
 			pod:                createPodWithCaps(&kapi.Capabilities{Add: []kapi.Capability{"foo"}}),
-			psps:               []*extensions.PodSecurityPolicy{restricted, requiresFooToBeDropped},
+			psps:               []*policy.PodSecurityPolicy{restricted, requiresFooToBeDropped},
 			shouldPassAdmit:    false,
 			shouldPassValidate: false,
 		},
@@ -564,7 +563,7 @@ func TestAdmitCaps(t *testing.T) {
 		// a manual request to drop the cap.
 		"should accept cap drop when cap is required to be dropped": {
 			pod:                createPodWithCaps(&kapi.Capabilities{Drop: []kapi.Capability{"foo"}}),
-			psps:               []*extensions.PodSecurityPolicy{requiresFooToBeDropped},
+			psps:               []*policy.PodSecurityPolicy{requiresFooToBeDropped},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPSP:        requiresFooToBeDropped.Name,
@@ -572,7 +571,7 @@ func TestAdmitCaps(t *testing.T) {
 		// UC 6: required add is defaulted
 		"required add is defaulted": {
 			pod:                goodPod(),
-			psps:               []*extensions.PodSecurityPolicy{allowsFooInRequired},
+			psps:               []*policy.PodSecurityPolicy{allowsFooInRequired},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedCapabilities: &kapi.Capabilities{
@@ -583,7 +582,7 @@ func TestAdmitCaps(t *testing.T) {
 		// UC 7: required drop is defaulted
 		"required drop is defaulted": {
 			pod:                goodPod(),
-			psps:               []*extensions.PodSecurityPolicy{requiresFooToBeDropped},
+			psps:               []*policy.PodSecurityPolicy{requiresFooToBeDropped},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedCapabilities: &kapi.Capabilities{
@@ -594,7 +593,7 @@ func TestAdmitCaps(t *testing.T) {
 		// UC 8: using '*' in allowed caps
 		"should accept cap add when all caps are allowed": {
 			pod:                createPodWithCaps(&kapi.Capabilities{Add: []kapi.Capability{"foo"}}),
-			psps:               []*extensions.PodSecurityPolicy{restricted, allowAllInAllowed},
+			psps:               []*policy.PodSecurityPolicy{restricted, allowAllInAllowed},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPSP:        allowAllInAllowed.Name,
@@ -652,19 +651,19 @@ func TestAdmitVolumes(t *testing.T) {
 		psp := restrictivePSP()
 
 		// expect a denial for this PSP
-		testPSPAdmit(fmt.Sprintf("%s denial", string(fsType)), []*extensions.PodSecurityPolicy{psp}, pod, false, false, "", t)
+		testPSPAdmit(fmt.Sprintf("%s denial", string(fsType)), []*policy.PodSecurityPolicy{psp}, pod, false, false, "", t)
 
 		// also expect a denial for this PSP if it's an init container
 		useInitContainers(pod)
-		testPSPAdmit(fmt.Sprintf("%s denial", string(fsType)), []*extensions.PodSecurityPolicy{psp}, pod, false, false, "", t)
+		testPSPAdmit(fmt.Sprintf("%s denial", string(fsType)), []*policy.PodSecurityPolicy{psp}, pod, false, false, "", t)
 
 		// now add the fstype directly to the psp and it should validate
-		psp.Spec.Volumes = []extensions.FSType{fsType}
-		testPSPAdmit(fmt.Sprintf("%s direct accept", string(fsType)), []*extensions.PodSecurityPolicy{psp}, pod, true, true, psp.Name, t)
+		psp.Spec.Volumes = []policy.FSType{fsType}
+		testPSPAdmit(fmt.Sprintf("%s direct accept", string(fsType)), []*policy.PodSecurityPolicy{psp}, pod, true, true, psp.Name, t)
 
 		// now change the psp to allow any volumes and the pod should still validate
-		psp.Spec.Volumes = []extensions.FSType{extensions.All}
-		testPSPAdmit(fmt.Sprintf("%s wildcard accept", string(fsType)), []*extensions.PodSecurityPolicy{psp}, pod, true, true, psp.Name, t)
+		psp.Spec.Volumes = []policy.FSType{policy.All}
+		testPSPAdmit(fmt.Sprintf("%s wildcard accept", string(fsType)), []*policy.PodSecurityPolicy{psp}, pod, true, true, psp.Name, t)
 	}
 }
 
@@ -685,7 +684,7 @@ func TestAdmitHostNetwork(t *testing.T) {
 
 	tests := map[string]struct {
 		pod                 *kapi.Pod
-		psps                []*extensions.PodSecurityPolicy
+		psps                []*policy.PodSecurityPolicy
 		shouldPassAdmit     bool
 		shouldPassValidate  bool
 		expectedHostNetwork bool
@@ -693,7 +692,7 @@ func TestAdmitHostNetwork(t *testing.T) {
 	}{
 		"pod without hostnetwork request allowed under noHostNetwork PSP": {
 			pod:                 goodPod(),
-			psps:                []*extensions.PodSecurityPolicy{noHostNetwork},
+			psps:                []*policy.PodSecurityPolicy{noHostNetwork},
 			shouldPassAdmit:     true,
 			shouldPassValidate:  true,
 			expectedHostNetwork: false,
@@ -701,7 +700,7 @@ func TestAdmitHostNetwork(t *testing.T) {
 		},
 		"pod without hostnetwork request allowed under hostNetwork PSP": {
 			pod:                 goodPod(),
-			psps:                []*extensions.PodSecurityPolicy{hostNetwork},
+			psps:                []*policy.PodSecurityPolicy{hostNetwork},
 			shouldPassAdmit:     true,
 			shouldPassValidate:  true,
 			expectedHostNetwork: false,
@@ -709,13 +708,13 @@ func TestAdmitHostNetwork(t *testing.T) {
 		},
 		"pod with hostnetwork request denied by noHostNetwork PSP": {
 			pod:                createPodWithHostNetwork(true),
-			psps:               []*extensions.PodSecurityPolicy{noHostNetwork},
+			psps:               []*policy.PodSecurityPolicy{noHostNetwork},
 			shouldPassAdmit:    false,
 			shouldPassValidate: false,
 		},
 		"pod with hostnetwork request allowed by hostNetwork PSP": {
 			pod:                 createPodWithHostNetwork(true),
-			psps:                []*extensions.PodSecurityPolicy{noHostNetwork, hostNetwork},
+			psps:                []*policy.PodSecurityPolicy{noHostNetwork, hostNetwork},
 			shouldPassAdmit:     true,
 			shouldPassValidate:  true,
 			expectedHostNetwork: true,
@@ -760,47 +759,47 @@ func TestAdmitHostPorts(t *testing.T) {
 
 	hostPorts := restrictivePSP()
 	hostPorts.Name = "hostPorts"
-	hostPorts.Spec.HostPorts = []extensions.HostPortRange{
+	hostPorts.Spec.HostPorts = []policy.HostPortRange{
 		{Min: 1, Max: 10},
 	}
 
 	tests := map[string]struct {
 		pod                *kapi.Pod
-		psps               []*extensions.PodSecurityPolicy
+		psps               []*policy.PodSecurityPolicy
 		shouldPassAdmit    bool
 		shouldPassValidate bool
 		expectedPSP        string
 	}{
 		"host port out of range": {
 			pod:                createPodWithHostPorts(11),
-			psps:               []*extensions.PodSecurityPolicy{hostPorts},
+			psps:               []*policy.PodSecurityPolicy{hostPorts},
 			shouldPassAdmit:    false,
 			shouldPassValidate: false,
 		},
 		"host port in range": {
 			pod:                createPodWithHostPorts(5),
-			psps:               []*extensions.PodSecurityPolicy{hostPorts},
+			psps:               []*policy.PodSecurityPolicy{hostPorts},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPSP:        hostPorts.Name,
 		},
 		"no host ports with range": {
 			pod:                goodPod(),
-			psps:               []*extensions.PodSecurityPolicy{hostPorts},
+			psps:               []*policy.PodSecurityPolicy{hostPorts},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPSP:        hostPorts.Name,
 		},
 		"no host ports without range": {
 			pod:                goodPod(),
-			psps:               []*extensions.PodSecurityPolicy{noHostPorts},
+			psps:               []*policy.PodSecurityPolicy{noHostPorts},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPSP:        noHostPorts.Name,
 		},
 		"host ports without range": {
 			pod:                createPodWithHostPorts(5),
-			psps:               []*extensions.PodSecurityPolicy{noHostPorts},
+			psps:               []*policy.PodSecurityPolicy{noHostPorts},
 			shouldPassAdmit:    false,
 			shouldPassValidate: false,
 		},
@@ -831,7 +830,7 @@ func TestAdmitHostPID(t *testing.T) {
 
 	tests := map[string]struct {
 		pod                *kapi.Pod
-		psps               []*extensions.PodSecurityPolicy
+		psps               []*policy.PodSecurityPolicy
 		shouldPassAdmit    bool
 		shouldPassValidate bool
 		expectedHostPID    bool
@@ -839,7 +838,7 @@ func TestAdmitHostPID(t *testing.T) {
 	}{
 		"pod without hostpid request allowed under noHostPID PSP": {
 			pod:                goodPod(),
-			psps:               []*extensions.PodSecurityPolicy{noHostPID},
+			psps:               []*policy.PodSecurityPolicy{noHostPID},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedHostPID:    false,
@@ -847,7 +846,7 @@ func TestAdmitHostPID(t *testing.T) {
 		},
 		"pod without hostpid request allowed under hostPID PSP": {
 			pod:                goodPod(),
-			psps:               []*extensions.PodSecurityPolicy{hostPID},
+			psps:               []*policy.PodSecurityPolicy{hostPID},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedHostPID:    false,
@@ -855,12 +854,12 @@ func TestAdmitHostPID(t *testing.T) {
 		},
 		"pod with hostpid request denied by noHostPID PSP": {
 			pod:             createPodWithHostPID(true),
-			psps:            []*extensions.PodSecurityPolicy{noHostPID},
+			psps:            []*policy.PodSecurityPolicy{noHostPID},
 			shouldPassAdmit: false,
 		},
 		"pod with hostpid request allowed by hostPID PSP": {
 			pod:                createPodWithHostPID(true),
-			psps:               []*extensions.PodSecurityPolicy{noHostPID, hostPID},
+			psps:               []*policy.PodSecurityPolicy{noHostPID, hostPID},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedHostPID:    true,
@@ -896,7 +895,7 @@ func TestAdmitHostIPC(t *testing.T) {
 
 	tests := map[string]struct {
 		pod                *kapi.Pod
-		psps               []*extensions.PodSecurityPolicy
+		psps               []*policy.PodSecurityPolicy
 		shouldPassAdmit    bool
 		shouldPassValidate bool
 		expectedHostIPC    bool
@@ -904,7 +903,7 @@ func TestAdmitHostIPC(t *testing.T) {
 	}{
 		"pod without hostIPC request allowed under noHostIPC PSP": {
 			pod:                goodPod(),
-			psps:               []*extensions.PodSecurityPolicy{noHostIPC},
+			psps:               []*policy.PodSecurityPolicy{noHostIPC},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedHostIPC:    false,
@@ -912,7 +911,7 @@ func TestAdmitHostIPC(t *testing.T) {
 		},
 		"pod without hostIPC request allowed under hostIPC PSP": {
 			pod:                goodPod(),
-			psps:               []*extensions.PodSecurityPolicy{hostIPC},
+			psps:               []*policy.PodSecurityPolicy{hostIPC},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedHostIPC:    false,
@@ -920,13 +919,13 @@ func TestAdmitHostIPC(t *testing.T) {
 		},
 		"pod with hostIPC request denied by noHostIPC PSP": {
 			pod:                createPodWithHostIPC(true),
-			psps:               []*extensions.PodSecurityPolicy{noHostIPC},
+			psps:               []*policy.PodSecurityPolicy{noHostIPC},
 			shouldPassAdmit:    false,
 			shouldPassValidate: false,
 		},
 		"pod with hostIPC request allowed by hostIPC PSP": {
 			pod:                createPodWithHostIPC(true),
-			psps:               []*extensions.PodSecurityPolicy{noHostIPC, hostIPC},
+			psps:               []*policy.PodSecurityPolicy{noHostIPC, hostIPC},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedHostIPC:    true,
@@ -955,12 +954,12 @@ func createPodWithSecurityContexts(podSC *kapi.PodSecurityContext, containerSC *
 func TestAdmitSELinux(t *testing.T) {
 	runAsAny := permissivePSP()
 	runAsAny.Name = "runAsAny"
-	runAsAny.Spec.SELinux.Rule = extensions.SELinuxStrategyRunAsAny
+	runAsAny.Spec.SELinux.Rule = policy.SELinuxStrategyRunAsAny
 	runAsAny.Spec.SELinux.SELinuxOptions = nil
 
 	mustRunAs := permissivePSP()
 	mustRunAs.Name = "mustRunAs"
-	mustRunAs.Spec.SELinux.Rule = extensions.SELinuxStrategyMustRunAs
+	mustRunAs.Spec.SELinux.Rule = policy.SELinuxStrategyMustRunAs
 	mustRunAs.Spec.SELinux.SELinuxOptions = &kapi.SELinuxOptions{}
 	mustRunAs.Spec.SELinux.SELinuxOptions.Level = "level"
 	mustRunAs.Spec.SELinux.SELinuxOptions.Role = "role"
@@ -969,7 +968,7 @@ func TestAdmitSELinux(t *testing.T) {
 
 	tests := map[string]struct {
 		pod                 *kapi.Pod
-		psps                []*extensions.PodSecurityPolicy
+		psps                []*policy.PodSecurityPolicy
 		shouldPassAdmit     bool
 		shouldPassValidate  bool
 		expectedPodSC       *kapi.PodSecurityContext
@@ -978,7 +977,7 @@ func TestAdmitSELinux(t *testing.T) {
 	}{
 		"runAsAny with no request": {
 			pod:                 createPodWithSecurityContexts(nil, nil),
-			psps:                []*extensions.PodSecurityPolicy{runAsAny},
+			psps:                []*policy.PodSecurityPolicy{runAsAny},
 			shouldPassAdmit:     true,
 			shouldPassValidate:  true,
 			expectedPodSC:       nil,
@@ -987,7 +986,7 @@ func TestAdmitSELinux(t *testing.T) {
 		},
 		"runAsAny with empty pod request": {
 			pod:                 createPodWithSecurityContexts(&kapi.PodSecurityContext{}, nil),
-			psps:                []*extensions.PodSecurityPolicy{runAsAny},
+			psps:                []*policy.PodSecurityPolicy{runAsAny},
 			shouldPassAdmit:     true,
 			shouldPassValidate:  true,
 			expectedPodSC:       &kapi.PodSecurityContext{},
@@ -996,7 +995,7 @@ func TestAdmitSELinux(t *testing.T) {
 		},
 		"runAsAny with empty container request": {
 			pod:                 createPodWithSecurityContexts(nil, &kapi.SecurityContext{}),
-			psps:                []*extensions.PodSecurityPolicy{runAsAny},
+			psps:                []*policy.PodSecurityPolicy{runAsAny},
 			shouldPassAdmit:     true,
 			shouldPassValidate:  true,
 			expectedPodSC:       nil,
@@ -1006,7 +1005,7 @@ func TestAdmitSELinux(t *testing.T) {
 
 		"runAsAny with pod request": {
 			pod:                 createPodWithSecurityContexts(&kapi.PodSecurityContext{SELinuxOptions: &kapi.SELinuxOptions{User: "foo"}}, nil),
-			psps:                []*extensions.PodSecurityPolicy{runAsAny},
+			psps:                []*policy.PodSecurityPolicy{runAsAny},
 			shouldPassAdmit:     true,
 			shouldPassValidate:  true,
 			expectedPodSC:       &kapi.PodSecurityContext{SELinuxOptions: &kapi.SELinuxOptions{User: "foo"}},
@@ -1015,7 +1014,7 @@ func TestAdmitSELinux(t *testing.T) {
 		},
 		"runAsAny with container request": {
 			pod:                 createPodWithSecurityContexts(nil, &kapi.SecurityContext{SELinuxOptions: &kapi.SELinuxOptions{User: "foo"}}),
-			psps:                []*extensions.PodSecurityPolicy{runAsAny},
+			psps:                []*policy.PodSecurityPolicy{runAsAny},
 			shouldPassAdmit:     true,
 			shouldPassValidate:  true,
 			expectedPodSC:       nil,
@@ -1024,7 +1023,7 @@ func TestAdmitSELinux(t *testing.T) {
 		},
 		"runAsAny with pod and container request": {
 			pod:                 createPodWithSecurityContexts(&kapi.PodSecurityContext{SELinuxOptions: &kapi.SELinuxOptions{User: "bar"}}, &kapi.SecurityContext{SELinuxOptions: &kapi.SELinuxOptions{User: "foo"}}),
-			psps:                []*extensions.PodSecurityPolicy{runAsAny},
+			psps:                []*policy.PodSecurityPolicy{runAsAny},
 			shouldPassAdmit:     true,
 			shouldPassValidate:  true,
 			expectedPodSC:       &kapi.PodSecurityContext{SELinuxOptions: &kapi.SELinuxOptions{User: "bar"}},
@@ -1034,19 +1033,19 @@ func TestAdmitSELinux(t *testing.T) {
 
 		"mustRunAs with bad pod request": {
 			pod:                createPodWithSecurityContexts(&kapi.PodSecurityContext{SELinuxOptions: &kapi.SELinuxOptions{User: "foo"}}, nil),
-			psps:               []*extensions.PodSecurityPolicy{mustRunAs},
+			psps:               []*policy.PodSecurityPolicy{mustRunAs},
 			shouldPassAdmit:    false,
 			shouldPassValidate: false,
 		},
 		"mustRunAs with bad container request": {
 			pod:                createPodWithSecurityContexts(nil, &kapi.SecurityContext{SELinuxOptions: &kapi.SELinuxOptions{User: "foo"}}),
-			psps:               []*extensions.PodSecurityPolicy{mustRunAs},
+			psps:               []*policy.PodSecurityPolicy{mustRunAs},
 			shouldPassAdmit:    false,
 			shouldPassValidate: false,
 		},
 		"mustRunAs with no request": {
 			pod:                 createPodWithSecurityContexts(nil, nil),
-			psps:                []*extensions.PodSecurityPolicy{mustRunAs},
+			psps:                []*policy.PodSecurityPolicy{mustRunAs},
 			shouldPassAdmit:     true,
 			shouldPassValidate:  true,
 			expectedPodSC:       &kapi.PodSecurityContext{SELinuxOptions: mustRunAs.Spec.SELinux.SELinuxOptions},
@@ -1058,7 +1057,7 @@ func TestAdmitSELinux(t *testing.T) {
 				&kapi.PodSecurityContext{SELinuxOptions: &kapi.SELinuxOptions{Level: "level", Role: "role", Type: "type", User: "user"}},
 				nil,
 			),
-			psps:                []*extensions.PodSecurityPolicy{mustRunAs},
+			psps:                []*policy.PodSecurityPolicy{mustRunAs},
 			shouldPassAdmit:     true,
 			shouldPassValidate:  true,
 			expectedPodSC:       &kapi.PodSecurityContext{SELinuxOptions: mustRunAs.Spec.SELinux.SELinuxOptions},
@@ -1070,7 +1069,7 @@ func TestAdmitSELinux(t *testing.T) {
 				&kapi.PodSecurityContext{SELinuxOptions: &kapi.SELinuxOptions{Level: "level", Role: "role", Type: "type", User: "user"}},
 				nil,
 			),
-			psps:                []*extensions.PodSecurityPolicy{mustRunAs},
+			psps:                []*policy.PodSecurityPolicy{mustRunAs},
 			shouldPassAdmit:     true,
 			shouldPassValidate:  true,
 			expectedPodSC:       &kapi.PodSecurityContext{SELinuxOptions: mustRunAs.Spec.SELinux.SELinuxOptions},
@@ -1117,7 +1116,7 @@ func TestAdmitAppArmor(t *testing.T) {
 
 	tests := map[string]struct {
 		pod                *kapi.Pod
-		psp                *extensions.PodSecurityPolicy
+		psp                *policy.PodSecurityPolicy
 		shouldPassAdmit    bool
 		shouldPassValidate bool
 		expectedProfile    string
@@ -1172,7 +1171,7 @@ func TestAdmitAppArmor(t *testing.T) {
 	}
 
 	for k, v := range tests {
-		testPSPAdmit(k, []*extensions.PodSecurityPolicy{v.psp}, v.pod, v.shouldPassAdmit, v.shouldPassValidate, v.psp.Name, t)
+		testPSPAdmit(k, []*policy.PodSecurityPolicy{v.psp}, v.pod, v.shouldPassAdmit, v.shouldPassValidate, v.psp.Name, t)
 
 		if v.shouldPassAdmit {
 			assert.Equal(t, v.expectedProfile, apparmor.GetProfileNameFromPodAnnotations(v.pod.Annotations, defaultContainerName), k)
@@ -1190,24 +1189,24 @@ func TestAdmitRunAsUser(t *testing.T) {
 
 	runAsAny := permissivePSP()
 	runAsAny.Name = "runAsAny"
-	runAsAny.Spec.RunAsUser.Rule = extensions.RunAsUserStrategyRunAsAny
+	runAsAny.Spec.RunAsUser.Rule = policy.RunAsUserStrategyRunAsAny
 
 	mustRunAs := permissivePSP()
 	mustRunAs.Name = "mustRunAs"
-	mustRunAs.Spec.RunAsUser.Rule = extensions.RunAsUserStrategyMustRunAs
-	mustRunAs.Spec.RunAsUser.Ranges = []extensions.UserIDRange{
+	mustRunAs.Spec.RunAsUser.Rule = policy.RunAsUserStrategyMustRunAs
+	mustRunAs.Spec.RunAsUser.Ranges = []policy.UserIDRange{
 		{Min: int64(999), Max: int64(1000)},
 	}
 
 	runAsNonRoot := permissivePSP()
 	runAsNonRoot.Name = "runAsNonRoot"
-	runAsNonRoot.Spec.RunAsUser.Rule = extensions.RunAsUserStrategyMustRunAsNonRoot
+	runAsNonRoot.Spec.RunAsUser.Rule = policy.RunAsUserStrategyMustRunAsNonRoot
 
 	trueValue := true
 
 	tests := map[string]struct {
 		pod                 *kapi.Pod
-		psps                []*extensions.PodSecurityPolicy
+		psps                []*policy.PodSecurityPolicy
 		shouldPassAdmit     bool
 		shouldPassValidate  bool
 		expectedPodSC       *kapi.PodSecurityContext
@@ -1216,7 +1215,7 @@ func TestAdmitRunAsUser(t *testing.T) {
 	}{
 		"runAsAny no pod request": {
 			pod:                 createPodWithSecurityContexts(nil, nil),
-			psps:                []*extensions.PodSecurityPolicy{runAsAny},
+			psps:                []*policy.PodSecurityPolicy{runAsAny},
 			shouldPassAdmit:     true,
 			shouldPassValidate:  true,
 			expectedPodSC:       nil,
@@ -1225,7 +1224,7 @@ func TestAdmitRunAsUser(t *testing.T) {
 		},
 		"runAsAny pod request": {
 			pod:                 createPodWithSecurityContexts(podSC(utilpointer.Int64Ptr(1)), nil),
-			psps:                []*extensions.PodSecurityPolicy{runAsAny},
+			psps:                []*policy.PodSecurityPolicy{runAsAny},
 			shouldPassAdmit:     true,
 			shouldPassValidate:  true,
 			expectedPodSC:       podSC(utilpointer.Int64Ptr(1)),
@@ -1234,7 +1233,7 @@ func TestAdmitRunAsUser(t *testing.T) {
 		},
 		"runAsAny container request": {
 			pod:                 createPodWithSecurityContexts(nil, containerSC(utilpointer.Int64Ptr(1))),
-			psps:                []*extensions.PodSecurityPolicy{runAsAny},
+			psps:                []*policy.PodSecurityPolicy{runAsAny},
 			shouldPassAdmit:     true,
 			shouldPassValidate:  true,
 			expectedPodSC:       nil,
@@ -1244,20 +1243,20 @@ func TestAdmitRunAsUser(t *testing.T) {
 
 		"mustRunAs pod request out of range": {
 			pod:                createPodWithSecurityContexts(podSC(utilpointer.Int64Ptr(1)), nil),
-			psps:               []*extensions.PodSecurityPolicy{mustRunAs},
+			psps:               []*policy.PodSecurityPolicy{mustRunAs},
 			shouldPassAdmit:    false,
 			shouldPassValidate: false,
 		},
 		"mustRunAs container request out of range": {
 			pod:                createPodWithSecurityContexts(podSC(utilpointer.Int64Ptr(999)), containerSC(utilpointer.Int64Ptr(1))),
-			psps:               []*extensions.PodSecurityPolicy{mustRunAs},
+			psps:               []*policy.PodSecurityPolicy{mustRunAs},
 			shouldPassAdmit:    false,
 			shouldPassValidate: false,
 		},
 
 		"mustRunAs pod request in range": {
 			pod:                 createPodWithSecurityContexts(podSC(utilpointer.Int64Ptr(999)), nil),
-			psps:                []*extensions.PodSecurityPolicy{mustRunAs},
+			psps:                []*policy.PodSecurityPolicy{mustRunAs},
 			shouldPassAdmit:     true,
 			shouldPassValidate:  true,
 			expectedPodSC:       podSC(&mustRunAs.Spec.RunAsUser.Ranges[0].Min),
@@ -1266,7 +1265,7 @@ func TestAdmitRunAsUser(t *testing.T) {
 		},
 		"mustRunAs container request in range": {
 			pod:                 createPodWithSecurityContexts(nil, containerSC(utilpointer.Int64Ptr(999))),
-			psps:                []*extensions.PodSecurityPolicy{mustRunAs},
+			psps:                []*policy.PodSecurityPolicy{mustRunAs},
 			shouldPassAdmit:     true,
 			shouldPassValidate:  true,
 			expectedPodSC:       nil,
@@ -1275,7 +1274,7 @@ func TestAdmitRunAsUser(t *testing.T) {
 		},
 		"mustRunAs pod and container request in range": {
 			pod:                 createPodWithSecurityContexts(podSC(utilpointer.Int64Ptr(999)), containerSC(utilpointer.Int64Ptr(1000))),
-			psps:                []*extensions.PodSecurityPolicy{mustRunAs},
+			psps:                []*policy.PodSecurityPolicy{mustRunAs},
 			shouldPassAdmit:     true,
 			shouldPassValidate:  true,
 			expectedPodSC:       podSC(utilpointer.Int64Ptr(999)),
@@ -1284,7 +1283,7 @@ func TestAdmitRunAsUser(t *testing.T) {
 		},
 		"mustRunAs no request": {
 			pod:                 createPodWithSecurityContexts(nil, nil),
-			psps:                []*extensions.PodSecurityPolicy{mustRunAs},
+			psps:                []*policy.PodSecurityPolicy{mustRunAs},
 			shouldPassAdmit:     true,
 			shouldPassValidate:  true,
 			expectedPodSC:       nil,
@@ -1294,7 +1293,7 @@ func TestAdmitRunAsUser(t *testing.T) {
 
 		"runAsNonRoot no request": {
 			pod:                 createPodWithSecurityContexts(nil, nil),
-			psps:                []*extensions.PodSecurityPolicy{runAsNonRoot},
+			psps:                []*policy.PodSecurityPolicy{runAsNonRoot},
 			shouldPassAdmit:     true,
 			shouldPassValidate:  true,
 			expectedPodSC:       nil,
@@ -1303,13 +1302,13 @@ func TestAdmitRunAsUser(t *testing.T) {
 		},
 		"runAsNonRoot pod request root": {
 			pod:                createPodWithSecurityContexts(podSC(utilpointer.Int64Ptr(0)), nil),
-			psps:               []*extensions.PodSecurityPolicy{runAsNonRoot},
+			psps:               []*policy.PodSecurityPolicy{runAsNonRoot},
 			shouldPassAdmit:    false,
 			shouldPassValidate: false,
 		},
 		"runAsNonRoot pod request non-root": {
 			pod:                createPodWithSecurityContexts(podSC(utilpointer.Int64Ptr(1)), nil),
-			psps:               []*extensions.PodSecurityPolicy{runAsNonRoot},
+			psps:               []*policy.PodSecurityPolicy{runAsNonRoot},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPodSC:      podSC(utilpointer.Int64Ptr(1)),
@@ -1317,13 +1316,13 @@ func TestAdmitRunAsUser(t *testing.T) {
 		},
 		"runAsNonRoot container request root": {
 			pod:                createPodWithSecurityContexts(podSC(utilpointer.Int64Ptr(1)), containerSC(utilpointer.Int64Ptr(0))),
-			psps:               []*extensions.PodSecurityPolicy{runAsNonRoot},
+			psps:               []*policy.PodSecurityPolicy{runAsNonRoot},
 			shouldPassAdmit:    false,
 			shouldPassValidate: false,
 		},
 		"runAsNonRoot container request non-root": {
 			pod:                 createPodWithSecurityContexts(podSC(utilpointer.Int64Ptr(1)), containerSC(utilpointer.Int64Ptr(2))),
-			psps:                []*extensions.PodSecurityPolicy{runAsNonRoot},
+			psps:                []*policy.PodSecurityPolicy{runAsNonRoot},
 			shouldPassAdmit:     true,
 			shouldPassValidate:  true,
 			expectedPodSC:       podSC(utilpointer.Int64Ptr(1)),
@@ -1353,16 +1352,16 @@ func TestAdmitSupplementalGroups(t *testing.T) {
 
 	runAsAny := permissivePSP()
 	runAsAny.Name = "runAsAny"
-	runAsAny.Spec.SupplementalGroups.Rule = extensions.SupplementalGroupsStrategyRunAsAny
+	runAsAny.Spec.SupplementalGroups.Rule = policy.SupplementalGroupsStrategyRunAsAny
 
 	mustRunAs := permissivePSP()
 	mustRunAs.Name = "mustRunAs"
-	mustRunAs.Spec.SupplementalGroups.Rule = extensions.SupplementalGroupsStrategyMustRunAs
-	mustRunAs.Spec.SupplementalGroups.Ranges = []extensions.GroupIDRange{{Min: int64(999), Max: int64(1000)}}
+	mustRunAs.Spec.SupplementalGroups.Rule = policy.SupplementalGroupsStrategyMustRunAs
+	mustRunAs.Spec.SupplementalGroups.Ranges = []policy.GroupIDRange{{Min: int64(999), Max: int64(1000)}}
 
 	tests := map[string]struct {
 		pod                *kapi.Pod
-		psps               []*extensions.PodSecurityPolicy
+		psps               []*policy.PodSecurityPolicy
 		shouldPassAdmit    bool
 		shouldPassValidate bool
 		expectedPodSC      *kapi.PodSecurityContext
@@ -1370,7 +1369,7 @@ func TestAdmitSupplementalGroups(t *testing.T) {
 	}{
 		"runAsAny no pod request": {
 			pod:                createPodWithSecurityContexts(nil, nil),
-			psps:               []*extensions.PodSecurityPolicy{runAsAny},
+			psps:               []*policy.PodSecurityPolicy{runAsAny},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPodSC:      nil,
@@ -1378,7 +1377,7 @@ func TestAdmitSupplementalGroups(t *testing.T) {
 		},
 		"runAsAny empty pod request": {
 			pod:                createPodWithSecurityContexts(&kapi.PodSecurityContext{}, nil),
-			psps:               []*extensions.PodSecurityPolicy{runAsAny},
+			psps:               []*policy.PodSecurityPolicy{runAsAny},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPodSC:      &kapi.PodSecurityContext{},
@@ -1386,7 +1385,7 @@ func TestAdmitSupplementalGroups(t *testing.T) {
 		},
 		"runAsAny empty pod request empty supplemental groups": {
 			pod:                createPodWithSecurityContexts(&kapi.PodSecurityContext{SupplementalGroups: []int64{}}, nil),
-			psps:               []*extensions.PodSecurityPolicy{runAsAny},
+			psps:               []*policy.PodSecurityPolicy{runAsAny},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPodSC:      &kapi.PodSecurityContext{SupplementalGroups: []int64{}},
@@ -1394,7 +1393,7 @@ func TestAdmitSupplementalGroups(t *testing.T) {
 		},
 		"runAsAny pod request": {
 			pod:                createPodWithSecurityContexts(podSC(1), nil),
-			psps:               []*extensions.PodSecurityPolicy{runAsAny},
+			psps:               []*policy.PodSecurityPolicy{runAsAny},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPodSC:      &kapi.PodSecurityContext{SupplementalGroups: []int64{1}},
@@ -1402,7 +1401,7 @@ func TestAdmitSupplementalGroups(t *testing.T) {
 		},
 		"mustRunAs no pod request": {
 			pod:                createPodWithSecurityContexts(nil, nil),
-			psps:               []*extensions.PodSecurityPolicy{mustRunAs},
+			psps:               []*policy.PodSecurityPolicy{mustRunAs},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPodSC:      podSC(mustRunAs.Spec.SupplementalGroups.Ranges[0].Min),
@@ -1410,13 +1409,13 @@ func TestAdmitSupplementalGroups(t *testing.T) {
 		},
 		"mustRunAs bad pod request": {
 			pod:                createPodWithSecurityContexts(podSC(1), nil),
-			psps:               []*extensions.PodSecurityPolicy{mustRunAs},
+			psps:               []*policy.PodSecurityPolicy{mustRunAs},
 			shouldPassAdmit:    false,
 			shouldPassValidate: false,
 		},
 		"mustRunAs good pod request": {
 			pod:                createPodWithSecurityContexts(podSC(999), nil),
-			psps:               []*extensions.PodSecurityPolicy{mustRunAs},
+			psps:               []*policy.PodSecurityPolicy{mustRunAs},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPodSC:      podSC(999),
@@ -1447,14 +1446,14 @@ func TestAdmitFSGroup(t *testing.T) {
 
 	runAsAny := restrictivePSP()
 	runAsAny.Name = "runAsAny"
-	runAsAny.Spec.FSGroup.Rule = extensions.FSGroupStrategyRunAsAny
+	runAsAny.Spec.FSGroup.Rule = policy.FSGroupStrategyRunAsAny
 
 	mustRunAs := restrictivePSP()
 	mustRunAs.Name = "mustRunAs"
 
 	tests := map[string]struct {
 		pod                *kapi.Pod
-		psps               []*extensions.PodSecurityPolicy
+		psps               []*policy.PodSecurityPolicy
 		shouldPassAdmit    bool
 		shouldPassValidate bool
 		expectedFSGroup    *int64
@@ -1462,7 +1461,7 @@ func TestAdmitFSGroup(t *testing.T) {
 	}{
 		"runAsAny no pod request": {
 			pod:                goodPod(),
-			psps:               []*extensions.PodSecurityPolicy{runAsAny},
+			psps:               []*policy.PodSecurityPolicy{runAsAny},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedFSGroup:    nil,
@@ -1470,7 +1469,7 @@ func TestAdmitFSGroup(t *testing.T) {
 		},
 		"runAsAny pod request": {
 			pod:                createPodWithFSGroup(1),
-			psps:               []*extensions.PodSecurityPolicy{runAsAny},
+			psps:               []*policy.PodSecurityPolicy{runAsAny},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedFSGroup:    utilpointer.Int64Ptr(1),
@@ -1478,7 +1477,7 @@ func TestAdmitFSGroup(t *testing.T) {
 		},
 		"mustRunAs no pod request": {
 			pod:                goodPod(),
-			psps:               []*extensions.PodSecurityPolicy{mustRunAs},
+			psps:               []*policy.PodSecurityPolicy{mustRunAs},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedFSGroup:    &mustRunAs.Spec.SupplementalGroups.Ranges[0].Min,
@@ -1486,13 +1485,13 @@ func TestAdmitFSGroup(t *testing.T) {
 		},
 		"mustRunAs bad pod request": {
 			pod:                createPodWithFSGroup(1),
-			psps:               []*extensions.PodSecurityPolicy{mustRunAs},
+			psps:               []*policy.PodSecurityPolicy{mustRunAs},
 			shouldPassAdmit:    false,
 			shouldPassValidate: false,
 		},
 		"mustRunAs good pod request": {
 			pod:                createPodWithFSGroup(999),
-			psps:               []*extensions.PodSecurityPolicy{mustRunAs},
+			psps:               []*policy.PodSecurityPolicy{mustRunAs},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedFSGroup:    utilpointer.Int64Ptr(999),
@@ -1540,7 +1539,7 @@ func TestAdmitReadOnlyRootFilesystem(t *testing.T) {
 
 	tests := map[string]struct {
 		pod                *kapi.Pod
-		psps               []*extensions.PodSecurityPolicy
+		psps               []*policy.PodSecurityPolicy
 		shouldPassAdmit    bool
 		shouldPassValidate bool
 		expectedRORFS      bool
@@ -1548,7 +1547,7 @@ func TestAdmitReadOnlyRootFilesystem(t *testing.T) {
 	}{
 		"no-rorfs allows pod request with rorfs": {
 			pod:                createPodWithRORFS(true),
-			psps:               []*extensions.PodSecurityPolicy{noRORFS},
+			psps:               []*policy.PodSecurityPolicy{noRORFS},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedRORFS:      true,
@@ -1556,7 +1555,7 @@ func TestAdmitReadOnlyRootFilesystem(t *testing.T) {
 		},
 		"no-rorfs allows pod request without rorfs": {
 			pod:                createPodWithRORFS(false),
-			psps:               []*extensions.PodSecurityPolicy{noRORFS},
+			psps:               []*policy.PodSecurityPolicy{noRORFS},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedRORFS:      false,
@@ -1564,13 +1563,13 @@ func TestAdmitReadOnlyRootFilesystem(t *testing.T) {
 		},
 		"rorfs rejects pod request without rorfs": {
 			pod:                createPodWithRORFS(false),
-			psps:               []*extensions.PodSecurityPolicy{rorfs},
+			psps:               []*policy.PodSecurityPolicy{rorfs},
 			shouldPassAdmit:    false,
 			shouldPassValidate: false,
 		},
 		"rorfs defaults nil pod request": {
 			pod:                goodPod(),
-			psps:               []*extensions.PodSecurityPolicy{rorfs},
+			psps:               []*policy.PodSecurityPolicy{rorfs},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedRORFS:      true,
@@ -1578,7 +1577,7 @@ func TestAdmitReadOnlyRootFilesystem(t *testing.T) {
 		},
 		"rorfs accepts pod request with rorfs": {
 			pod:                createPodWithRORFS(true),
-			psps:               []*extensions.PodSecurityPolicy{rorfs},
+			psps:               []*policy.PodSecurityPolicy{rorfs},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedRORFS:      true,
@@ -1619,156 +1618,156 @@ func TestAdmitSysctls(t *testing.T) {
 
 	emptySysctls := restrictivePSP()
 	emptySysctls.Name = "empty sysctls"
-	emptySysctls.Annotations[extensions.SysctlsPodSecurityPolicyAnnotationKey] = ""
+	emptySysctls.Annotations[policy.SysctlsPodSecurityPolicyAnnotationKey] = ""
 
 	mixedSysctls := restrictivePSP()
 	mixedSysctls.Name = "wildcard sysctls"
-	mixedSysctls.Annotations[extensions.SysctlsPodSecurityPolicyAnnotationKey] = "a.*,b.*,c,d.e.f"
+	mixedSysctls.Annotations[policy.SysctlsPodSecurityPolicyAnnotationKey] = "a.*,b.*,c,d.e.f"
 
 	aSysctl := restrictivePSP()
 	aSysctl.Name = "a sysctl"
-	aSysctl.Annotations[extensions.SysctlsPodSecurityPolicyAnnotationKey] = "a"
+	aSysctl.Annotations[policy.SysctlsPodSecurityPolicyAnnotationKey] = "a"
 
 	bSysctl := restrictivePSP()
 	bSysctl.Name = "b sysctl"
-	bSysctl.Annotations[extensions.SysctlsPodSecurityPolicyAnnotationKey] = "b"
+	bSysctl.Annotations[policy.SysctlsPodSecurityPolicyAnnotationKey] = "b"
 
 	cSysctl := restrictivePSP()
 	cSysctl.Name = "c sysctl"
-	cSysctl.Annotations[extensions.SysctlsPodSecurityPolicyAnnotationKey] = "c"
+	cSysctl.Annotations[policy.SysctlsPodSecurityPolicyAnnotationKey] = "c"
 
 	catchallSysctls := restrictivePSP()
 	catchallSysctls.Name = "catchall sysctl"
-	catchallSysctls.Annotations[extensions.SysctlsPodSecurityPolicyAnnotationKey] = "*"
+	catchallSysctls.Annotations[policy.SysctlsPodSecurityPolicyAnnotationKey] = "*"
 
 	tests := map[string]struct {
 		pod                *kapi.Pod
-		psps               []*extensions.PodSecurityPolicy
+		psps               []*policy.PodSecurityPolicy
 		shouldPassAdmit    bool
 		shouldPassValidate bool
 		expectedPSP        string
 	}{
 		"pod without unsafe sysctls request allowed under noSysctls PSP": {
 			pod:                goodPod(),
-			psps:               []*extensions.PodSecurityPolicy{noSysctls},
+			psps:               []*policy.PodSecurityPolicy{noSysctls},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPSP:        noSysctls.Name,
 		},
 		"pod without any sysctls request allowed under emptySysctls PSP": {
 			pod:                goodPod(),
-			psps:               []*extensions.PodSecurityPolicy{emptySysctls},
+			psps:               []*policy.PodSecurityPolicy{emptySysctls},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPSP:        emptySysctls.Name,
 		},
 		"pod with safe sysctls request allowed under noSysctls PSP": {
 			pod:                podWithSysctls([]string{"a", "b"}, []string{}),
-			psps:               []*extensions.PodSecurityPolicy{noSysctls},
+			psps:               []*policy.PodSecurityPolicy{noSysctls},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPSP:        noSysctls.Name,
 		},
 		"pod with unsafe sysctls request allowed under noSysctls PSP": {
 			pod:                podWithSysctls([]string{}, []string{"a", "b"}),
-			psps:               []*extensions.PodSecurityPolicy{noSysctls},
+			psps:               []*policy.PodSecurityPolicy{noSysctls},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPSP:        noSysctls.Name,
 		},
 		"pod with safe sysctls request disallowed under emptySysctls PSP": {
 			pod:                podWithSysctls([]string{"a", "b"}, []string{}),
-			psps:               []*extensions.PodSecurityPolicy{emptySysctls},
+			psps:               []*policy.PodSecurityPolicy{emptySysctls},
 			shouldPassAdmit:    false,
 			shouldPassValidate: false,
 		},
 		"pod with unsafe sysctls a, b request disallowed under aSysctls SCC": {
 			pod:                podWithSysctls([]string{}, []string{"a", "b"}),
-			psps:               []*extensions.PodSecurityPolicy{aSysctl},
+			psps:               []*policy.PodSecurityPolicy{aSysctl},
 			shouldPassAdmit:    false,
 			shouldPassValidate: false,
 		},
 		"pod with unsafe sysctls b request disallowed under aSysctls SCC": {
 			pod:                podWithSysctls([]string{}, []string{"b"}),
-			psps:               []*extensions.PodSecurityPolicy{aSysctl},
+			psps:               []*policy.PodSecurityPolicy{aSysctl},
 			shouldPassAdmit:    false,
 			shouldPassValidate: false,
 		},
 		"pod with unsafe sysctls a request allowed under aSysctls SCC": {
 			pod:                podWithSysctls([]string{}, []string{"a"}),
-			psps:               []*extensions.PodSecurityPolicy{aSysctl},
+			psps:               []*policy.PodSecurityPolicy{aSysctl},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPSP:        aSysctl.Name,
 		},
 		"pod with safe sysctls a, b request disallowed under aSysctls SCC": {
 			pod:                podWithSysctls([]string{"a", "b"}, []string{}),
-			psps:               []*extensions.PodSecurityPolicy{aSysctl},
+			psps:               []*policy.PodSecurityPolicy{aSysctl},
 			shouldPassAdmit:    false,
 			shouldPassValidate: false,
 		},
 		"pod with safe sysctls b request disallowed under aSysctls SCC": {
 			pod:                podWithSysctls([]string{"b"}, []string{}),
-			psps:               []*extensions.PodSecurityPolicy{aSysctl},
+			psps:               []*policy.PodSecurityPolicy{aSysctl},
 			shouldPassAdmit:    false,
 			shouldPassValidate: false,
 		},
 		"pod with safe sysctls a request allowed under aSysctls SCC": {
 			pod:                podWithSysctls([]string{"a"}, []string{}),
-			psps:               []*extensions.PodSecurityPolicy{aSysctl},
+			psps:               []*policy.PodSecurityPolicy{aSysctl},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPSP:        aSysctl.Name,
 		},
 		"pod with unsafe sysctls request disallowed under emptySysctls PSP": {
 			pod:                podWithSysctls([]string{}, []string{"a", "b"}),
-			psps:               []*extensions.PodSecurityPolicy{emptySysctls},
+			psps:               []*policy.PodSecurityPolicy{emptySysctls},
 			shouldPassAdmit:    false,
 			shouldPassValidate: false,
 		},
 		"pod with matching sysctls request allowed under mixedSysctls PSP": {
 			pod:                podWithSysctls([]string{"a.b", "b.c"}, []string{"c", "d.e.f"}),
-			psps:               []*extensions.PodSecurityPolicy{mixedSysctls},
+			psps:               []*policy.PodSecurityPolicy{mixedSysctls},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPSP:        mixedSysctls.Name,
 		},
 		"pod with not-matching unsafe sysctls request disallowed under mixedSysctls PSP": {
 			pod:                podWithSysctls([]string{"a.b", "b.c", "c", "d.e.f"}, []string{"e"}),
-			psps:               []*extensions.PodSecurityPolicy{mixedSysctls},
+			psps:               []*policy.PodSecurityPolicy{mixedSysctls},
 			shouldPassAdmit:    false,
 			shouldPassValidate: false,
 		},
 		"pod with not-matching safe sysctls request disallowed under mixedSysctls PSP": {
 			pod:                podWithSysctls([]string{"a.b", "b.c", "c", "d.e.f", "e"}, []string{}),
-			psps:               []*extensions.PodSecurityPolicy{mixedSysctls},
+			psps:               []*policy.PodSecurityPolicy{mixedSysctls},
 			shouldPassAdmit:    false,
 			shouldPassValidate: false,
 		},
 		"pod with sysctls request allowed under catchallSysctls PSP": {
 			pod:                podWithSysctls([]string{"e"}, []string{"f"}),
-			psps:               []*extensions.PodSecurityPolicy{catchallSysctls},
+			psps:               []*policy.PodSecurityPolicy{catchallSysctls},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPSP:        catchallSysctls.Name,
 		},
 		"pod with sysctls request allowed under catchallSysctls PSP, not under mixedSysctls or emptySysctls PSP": {
 			pod:                podWithSysctls([]string{"e"}, []string{"f"}),
-			psps:               []*extensions.PodSecurityPolicy{mixedSysctls, catchallSysctls, emptySysctls},
+			psps:               []*policy.PodSecurityPolicy{mixedSysctls, catchallSysctls, emptySysctls},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPSP:        catchallSysctls.Name,
 		},
 		"pod with safe c sysctl request allowed under cSysctl PSP, not under aSysctl or bSysctl PSP": {
 			pod:                podWithSysctls([]string{}, []string{"c"}),
-			psps:               []*extensions.PodSecurityPolicy{aSysctl, bSysctl, cSysctl},
+			psps:               []*policy.PodSecurityPolicy{aSysctl, bSysctl, cSysctl},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPSP:        cSysctl.Name,
 		},
 		"pod with unsafe c sysctl request allowed under cSysctl PSP, not under aSysctl or bSysctl PSP": {
 			pod:                podWithSysctls([]string{"c"}, []string{}),
-			psps:               []*extensions.PodSecurityPolicy{aSysctl, bSysctl, cSysctl},
+			psps:               []*policy.PodSecurityPolicy{aSysctl, bSysctl, cSysctl},
 			shouldPassAdmit:    true,
 			shouldPassValidate: true,
 			expectedPSP:        cSysctl.Name,
@@ -1795,11 +1794,11 @@ func TestAdmitSysctls(t *testing.T) {
 	}
 }
 
-func testPSPAdmit(testCaseName string, psps []*extensions.PodSecurityPolicy, pod *kapi.Pod, shouldPassAdmit, shouldPassValidate bool, expectedPSP string, t *testing.T) {
+func testPSPAdmit(testCaseName string, psps []*policy.PodSecurityPolicy, pod *kapi.Pod, shouldPassAdmit, shouldPassValidate bool, expectedPSP string, t *testing.T) {
 	testPSPAdmitAdvanced(testCaseName, kadmission.Create, psps, nil, &user.DefaultInfo{}, pod, nil, shouldPassAdmit, shouldPassValidate, true, expectedPSP, t)
 }
 
-func testPSPAdmitAdvanced(testCaseName string, op kadmission.Operation, psps []*extensions.PodSecurityPolicy, authz authorizer.Authorizer, userInfo user.Info, pod, oldPod *kapi.Pod, shouldPassAdmit, shouldPassValidate bool, canMutate bool, expectedPSP string, t *testing.T) {
+func testPSPAdmitAdvanced(testCaseName string, op kadmission.Operation, psps []*policy.PodSecurityPolicy, authz authorizer.Authorizer, userInfo user.Info, pod, oldPod *kapi.Pod, shouldPassAdmit, shouldPassValidate bool, canMutate bool, expectedPSP string, t *testing.T) {
 	originalPod := pod.DeepCopy()
 	plugin := NewTestAdmission(psps, authz)
 
@@ -1907,50 +1906,50 @@ func TestAssignSecurityContext(t *testing.T) {
 func TestCreateProvidersFromConstraints(t *testing.T) {
 	testCases := map[string]struct {
 		// use a generating function so we can test for non-mutation
-		psp         func() *extensions.PodSecurityPolicy
+		psp         func() *policy.PodSecurityPolicy
 		expectedErr string
 	}{
 		"valid psp": {
-			psp: func() *extensions.PodSecurityPolicy {
-				return &extensions.PodSecurityPolicy{
+			psp: func() *policy.PodSecurityPolicy {
+				return &policy.PodSecurityPolicy{
 					ObjectMeta: metav1.ObjectMeta{
 						Name: "valid psp",
 					},
-					Spec: extensions.PodSecurityPolicySpec{
-						SELinux: extensions.SELinuxStrategyOptions{
-							Rule: extensions.SELinuxStrategyRunAsAny,
+					Spec: policy.PodSecurityPolicySpec{
+						SELinux: policy.SELinuxStrategyOptions{
+							Rule: policy.SELinuxStrategyRunAsAny,
 						},
-						RunAsUser: extensions.RunAsUserStrategyOptions{
-							Rule: extensions.RunAsUserStrategyRunAsAny,
+						RunAsUser: policy.RunAsUserStrategyOptions{
+							Rule: policy.RunAsUserStrategyRunAsAny,
 						},
-						FSGroup: extensions.FSGroupStrategyOptions{
-							Rule: extensions.FSGroupStrategyRunAsAny,
+						FSGroup: policy.FSGroupStrategyOptions{
+							Rule: policy.FSGroupStrategyRunAsAny,
 						},
-						SupplementalGroups: extensions.SupplementalGroupsStrategyOptions{
-							Rule: extensions.SupplementalGroupsStrategyRunAsAny,
+						SupplementalGroups: policy.SupplementalGroupsStrategyOptions{
+							Rule: policy.SupplementalGroupsStrategyRunAsAny,
 						},
 					},
 				}
 			},
 		},
 		"bad psp strategy options": {
-			psp: func() *extensions.PodSecurityPolicy {
-				return &extensions.PodSecurityPolicy{
+			psp: func() *policy.PodSecurityPolicy {
+				return &policy.PodSecurityPolicy{
 					ObjectMeta: metav1.ObjectMeta{
 						Name: "bad psp user options",
 					},
-					Spec: extensions.PodSecurityPolicySpec{
-						SELinux: extensions.SELinuxStrategyOptions{
-							Rule: extensions.SELinuxStrategyRunAsAny,
+					Spec: policy.PodSecurityPolicySpec{
+						SELinux: policy.SELinuxStrategyOptions{
+							Rule: policy.SELinuxStrategyRunAsAny,
 						},
-						RunAsUser: extensions.RunAsUserStrategyOptions{
-							Rule: extensions.RunAsUserStrategyMustRunAs,
+						RunAsUser: policy.RunAsUserStrategyOptions{
+							Rule: policy.RunAsUserStrategyMustRunAs,
 						},
-						FSGroup: extensions.FSGroupStrategyOptions{
-							Rule: extensions.FSGroupStrategyRunAsAny,
+						FSGroup: policy.FSGroupStrategyOptions{
+							Rule: policy.FSGroupStrategyRunAsAny,
 						},
-						SupplementalGroups: extensions.SupplementalGroupsStrategyOptions{
-							Rule: extensions.SupplementalGroupsStrategyRunAsAny,
+						SupplementalGroups: policy.SupplementalGroupsStrategyOptions{
+							Rule: policy.SupplementalGroupsStrategyRunAsAny,
 						},
 					},
 				}
@@ -1966,7 +1965,7 @@ func TestCreateProvidersFromConstraints(t *testing.T) {
 		}
 
 		psp := v.psp()
-		_, errs := admit.createProvidersFromPolicies([]*extensions.PodSecurityPolicy{psp}, "namespace")
+		_, errs := admit.createProvidersFromPolicies([]*policy.PodSecurityPolicy{psp}, "namespace")
 
 		if !reflect.DeepEqual(psp, v.psp()) {
 			diff := diff.ObjectDiff(psp, v.psp())
@@ -1991,7 +1990,7 @@ func TestCreateProvidersFromConstraints(t *testing.T) {
 }
 
 func TestPolicyAuthorization(t *testing.T) {
-	policyWithName := func(name string) *extensions.PodSecurityPolicy {
+	policyWithName := func(name string) *policy.PodSecurityPolicy {
 		p := permissivePSP()
 		p.Name = name
 		return p
@@ -2002,7 +2001,7 @@ func TestPolicyAuthorization(t *testing.T) {
 		sa             string
 		ns             string
 		expectedPolicy string
-		inPolicies     []*extensions.PodSecurityPolicy
+		inPolicies     []*policy.PodSecurityPolicy
 		allowed        map[string]map[string]map[string]bool
 		allowedGroup   string
 	}{
@@ -2015,7 +2014,7 @@ func TestPolicyAuthorization(t *testing.T) {
 					"test": {"policy": true},
 				},
 			},
-			inPolicies:     []*extensions.PodSecurityPolicy{policyWithName("policy")},
+			inPolicies:     []*policy.PodSecurityPolicy{policyWithName("policy")},
 			expectedPolicy: "policy",
 		},
 		"policy allowed by sa (extensions API Group)": {
@@ -2027,7 +2026,7 @@ func TestPolicyAuthorization(t *testing.T) {
 					"test": {"policy": true},
 				},
 			},
-			inPolicies:     []*extensions.PodSecurityPolicy{policyWithName("policy")},
+			inPolicies:     []*policy.PodSecurityPolicy{policyWithName("policy")},
 			expectedPolicy: "policy",
 		},
 		"policy allowed by user (policy API Group)": {
@@ -2039,7 +2038,7 @@ func TestPolicyAuthorization(t *testing.T) {
 					"test": {"policy": true},
 				},
 			},
-			inPolicies:     []*extensions.PodSecurityPolicy{policyWithName("policy")},
+			inPolicies:     []*policy.PodSecurityPolicy{policyWithName("policy")},
 			expectedPolicy: "policy",
 			allowedGroup:   policy.GroupName,
 		},
@@ -2052,7 +2051,7 @@ func TestPolicyAuthorization(t *testing.T) {
 					"test": {"policy": true},
 				},
 			},
-			inPolicies:     []*extensions.PodSecurityPolicy{policyWithName("policy")},
+			inPolicies:     []*policy.PodSecurityPolicy{policyWithName("policy")},
 			expectedPolicy: "policy",
 			allowedGroup:   policy.GroupName,
 		},
@@ -2061,7 +2060,7 @@ func TestPolicyAuthorization(t *testing.T) {
 			sa:             "sa",
 			ns:             "test",
 			allowed:        map[string]map[string]map[string]bool{},
-			inPolicies:     []*extensions.PodSecurityPolicy{policyWithName("policy")},
+			inPolicies:     []*policy.PodSecurityPolicy{policyWithName("policy")},
 			expectedPolicy: "",
 		},
 		"multiple policies allowed": {
@@ -2080,7 +2079,7 @@ func TestPolicyAuthorization(t *testing.T) {
 					"other": {"policy7": true},
 				},
 			},
-			inPolicies: []*extensions.PodSecurityPolicy{
+			inPolicies: []*policy.PodSecurityPolicy{
 				// Prefix to force checking these policies first.
 				policyWithName("a_policy1"), // not allowed in this namespace
 				policyWithName("a_policy2"), // not allowed in this namespace
@@ -2104,7 +2103,7 @@ func TestPolicyAuthorization(t *testing.T) {
 					"test": {"policy2": true},
 				},
 			},
-			inPolicies: []*extensions.PodSecurityPolicy{
+			inPolicies: []*policy.PodSecurityPolicy{
 				policyWithName("policy1"),
 				policyWithName("policy2"),
 				policyWithName("policy3"),
@@ -2124,7 +2123,7 @@ func TestPolicyAuthorization(t *testing.T) {
 					"test": {"policy2": true},
 				},
 			},
-			inPolicies: []*extensions.PodSecurityPolicy{
+			inPolicies: []*policy.PodSecurityPolicy{
 				policyWithName("policy1"),
 				policyWithName("policy2"),
 				policyWithName("policy3"),
@@ -2144,7 +2143,7 @@ func TestPolicyAuthorization(t *testing.T) {
 					"test": {"policy2": true},
 				},
 			},
-			inPolicies: []*extensions.PodSecurityPolicy{
+			inPolicies: []*policy.PodSecurityPolicy{
 				policyWithName("policy1"),
 				policyWithName("policy2"),
 				policyWithName("policy3"),
@@ -2169,7 +2168,7 @@ func TestPolicyAuthorization(t *testing.T) {
 }
 
 func TestPolicyAuthorizationErrors(t *testing.T) {
-	policyWithName := func(name string) *extensions.PodSecurityPolicy {
+	policyWithName := func(name string) *policy.PodSecurityPolicy {
 		p := restrictivePSP()
 		p.Name = name
 		return p
@@ -2182,13 +2181,13 @@ func TestPolicyAuthorizationErrors(t *testing.T) {
 	)
 
 	tests := map[string]struct {
-		inPolicies           []*extensions.PodSecurityPolicy
+		inPolicies           []*policy.PodSecurityPolicy
 		allowed              map[string]map[string]map[string]bool
 		expectValidationErrs int
 	}{
 		"policies not allowed": {
 			allowed: map[string]map[string]map[string]bool{},
-			inPolicies: []*extensions.PodSecurityPolicy{
+			inPolicies: []*policy.PodSecurityPolicy{
 				policyWithName("policy1"),
 				policyWithName("policy2"),
 			},
@@ -2200,7 +2199,7 @@ func TestPolicyAuthorizationErrors(t *testing.T) {
 					"test": {"policy1": true},
 				},
 			},
-			inPolicies: []*extensions.PodSecurityPolicy{
+			inPolicies: []*policy.PodSecurityPolicy{
 				policyWithName("policy1"),
 				policyWithName("policy2"),
 			},
@@ -2212,7 +2211,7 @@ func TestPolicyAuthorizationErrors(t *testing.T) {
 					"test": {"policy2": true},
 				},
 			},
-			inPolicies: []*extensions.PodSecurityPolicy{
+			inPolicies: []*policy.PodSecurityPolicy{
 				policyWithName("policy1"),
 				policyWithName("policy2"),
 			},
@@ -2227,7 +2226,7 @@ func TestPolicyAuthorizationErrors(t *testing.T) {
 					"test": {"policy2": true},
 				},
 			},
-			inPolicies: []*extensions.PodSecurityPolicy{
+			inPolicies: []*policy.PodSecurityPolicy{
 				policyWithName("policy1"),
 				policyWithName("policy2"),
 			},
@@ -2254,26 +2253,26 @@ func TestPolicyAuthorizationErrors(t *testing.T) {
 }
 
 func TestPreferValidatedPSP(t *testing.T) {
-	restrictivePSPWithName := func(name string) *extensions.PodSecurityPolicy {
+	restrictivePSPWithName := func(name string) *policy.PodSecurityPolicy {
 		p := restrictivePSP()
 		p.Name = name
 		return p
 	}
 
-	permissivePSPWithName := func(name string) *extensions.PodSecurityPolicy {
+	permissivePSPWithName := func(name string) *policy.PodSecurityPolicy {
 		p := permissivePSP()
 		p.Name = name
 		return p
 	}
 
 	tests := map[string]struct {
-		inPolicies           []*extensions.PodSecurityPolicy
+		inPolicies           []*policy.PodSecurityPolicy
 		expectValidationErrs int
 		validatedPSPHint     string
 		expectedPSP          string
 	}{
 		"no policy saved in annotations, PSPs are ordered lexicographically": {
-			inPolicies: []*extensions.PodSecurityPolicy{
+			inPolicies: []*policy.PodSecurityPolicy{
 				restrictivePSPWithName("001restrictive"),
 				restrictivePSPWithName("002restrictive"),
 				permissivePSPWithName("002permissive"),
@@ -2285,7 +2284,7 @@ func TestPreferValidatedPSP(t *testing.T) {
 			expectedPSP:          "001permissive",
 		},
 		"policy saved in annotations is preferred": {
-			inPolicies: []*extensions.PodSecurityPolicy{
+			inPolicies: []*policy.PodSecurityPolicy{
 				restrictivePSPWithName("001restrictive"),
 				restrictivePSPWithName("002restrictive"),
 				permissivePSPWithName("001permissive"),
@@ -2297,7 +2296,7 @@ func TestPreferValidatedPSP(t *testing.T) {
 			expectedPSP:          "002permissive",
 		},
 		"policy saved in annotations is invalid": {
-			inPolicies: []*extensions.PodSecurityPolicy{
+			inPolicies: []*policy.PodSecurityPolicy{
 				restrictivePSPWithName("001restrictive"),
 				restrictivePSPWithName("002restrictive"),
 			},
@@ -2306,7 +2305,7 @@ func TestPreferValidatedPSP(t *testing.T) {
 			expectedPSP:          "",
 		},
 		"policy saved in annotations is disallowed anymore": {
-			inPolicies: []*extensions.PodSecurityPolicy{
+			inPolicies: []*policy.PodSecurityPolicy{
 				restrictivePSPWithName("001restrictive"),
 				restrictivePSPWithName("002restrictive"),
 			},
@@ -2315,7 +2314,7 @@ func TestPreferValidatedPSP(t *testing.T) {
 			expectedPSP:          "",
 		},
 		"policy saved in annotations is disallowed anymore, but find another one": {
-			inPolicies: []*extensions.PodSecurityPolicy{
+			inPolicies: []*policy.PodSecurityPolicy{
 				restrictivePSPWithName("001restrictive"),
 				restrictivePSPWithName("002restrictive"),
 				permissivePSPWithName("002permissive"),
@@ -2346,34 +2345,34 @@ func TestPreferValidatedPSP(t *testing.T) {
 	}
 }
 
-func restrictivePSP() *extensions.PodSecurityPolicy {
-	return &extensions.PodSecurityPolicy{
+func restrictivePSP() *policy.PodSecurityPolicy {
+	return &policy.PodSecurityPolicy{
 		ObjectMeta: metav1.ObjectMeta{
 			Name:        "restrictive",
 			Annotations: map[string]string{},
 		},
-		Spec: extensions.PodSecurityPolicySpec{
-			RunAsUser: extensions.RunAsUserStrategyOptions{
-				Rule: extensions.RunAsUserStrategyMustRunAs,
-				Ranges: []extensions.UserIDRange{
+		Spec: policy.PodSecurityPolicySpec{
+			RunAsUser: policy.RunAsUserStrategyOptions{
+				Rule: policy.RunAsUserStrategyMustRunAs,
+				Ranges: []policy.UserIDRange{
 					{Min: int64(999), Max: int64(999)},
 				},
 			},
-			SELinux: extensions.SELinuxStrategyOptions{
-				Rule: extensions.SELinuxStrategyMustRunAs,
+			SELinux: policy.SELinuxStrategyOptions{
+				Rule: policy.SELinuxStrategyMustRunAs,
 				SELinuxOptions: &kapi.SELinuxOptions{
 					Level: "s9:z0,z1",
 				},
 			},
-			FSGroup: extensions.FSGroupStrategyOptions{
-				Rule: extensions.FSGroupStrategyMustRunAs,
-				Ranges: []extensions.GroupIDRange{
+			FSGroup: policy.FSGroupStrategyOptions{
+				Rule: policy.FSGroupStrategyMustRunAs,
+				Ranges: []policy.GroupIDRange{
 					{Min: int64(999), Max: int64(999)},
 				},
 			},
-			SupplementalGroups: extensions.SupplementalGroupsStrategyOptions{
-				Rule: extensions.SupplementalGroupsStrategyMustRunAs,
-				Ranges: []extensions.GroupIDRange{
+			SupplementalGroups: policy.SupplementalGroupsStrategyOptions{
+				Rule: policy.SupplementalGroupsStrategyMustRunAs,
+				Ranges: []policy.GroupIDRange{
 					{Min: int64(999), Max: int64(999)},
 				},
 			},
@@ -2381,31 +2380,31 @@ func restrictivePSP() *extensions.PodSecurityPolicy {
 	}
 }
 
-func permissivePSP() *extensions.PodSecurityPolicy {
-	return &extensions.PodSecurityPolicy{
+func permissivePSP() *policy.PodSecurityPolicy {
+	return &policy.PodSecurityPolicy{
 		ObjectMeta: metav1.ObjectMeta{
 			Name:        "privileged",
 			Annotations: map[string]string{},
 		},
-		Spec: extensions.PodSecurityPolicySpec{
+		Spec: policy.PodSecurityPolicySpec{
 			AllowPrivilegeEscalation: true,
 			HostIPC:                  true,
 			HostNetwork:              true,
 			HostPID:                  true,
-			HostPorts:                []extensions.HostPortRange{{Min: 0, Max: 65536}},
-			Volumes:                  []extensions.FSType{extensions.All},
-			AllowedCapabilities:      []kapi.Capability{extensions.AllowAllCapabilities},
-			RunAsUser: extensions.RunAsUserStrategyOptions{
-				Rule: extensions.RunAsUserStrategyRunAsAny,
+			HostPorts:                []policy.HostPortRange{{Min: 0, Max: 65536}},
+			Volumes:                  []policy.FSType{policy.All},
+			AllowedCapabilities:      []kapi.Capability{policy.AllowAllCapabilities},
+			RunAsUser: policy.RunAsUserStrategyOptions{
+				Rule: policy.RunAsUserStrategyRunAsAny,
 			},
-			SELinux: extensions.SELinuxStrategyOptions{
-				Rule: extensions.SELinuxStrategyRunAsAny,
+			SELinux: policy.SELinuxStrategyOptions{
+				Rule: policy.SELinuxStrategyRunAsAny,
 			},
-			FSGroup: extensions.FSGroupStrategyOptions{
-				Rule: extensions.FSGroupStrategyRunAsAny,
+			FSGroup: policy.FSGroupStrategyOptions{
+				Rule: policy.FSGroupStrategyRunAsAny,
 			},
-			SupplementalGroups: extensions.SupplementalGroupsStrategyOptions{
-				Rule: extensions.SupplementalGroupsStrategyRunAsAny,
+			SupplementalGroups: policy.SupplementalGroupsStrategyOptions{
+				Rule: policy.SupplementalGroupsStrategyRunAsAny,
 			},
 		},
 	}
